chore: add common crate for project wide definitions
