+++ common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "zk_evm_common"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+ethereum-types = { workspace = true }
